rename to clip and remove more unnecessary files
