+++ clip.py
