+++ generate.py
@@ -3,7 +3,7 @@ import torch
 from PIL import Image
 import matplotlib.pyplot as plt
 from transformers import GPT2Tokenizer
-from clip_model_clean import load_clip
+from clip import load_clip
 from vision_language_connector import VisionLanguageConnector
 from gpt import GPT, GPTConfig, transpose_specific_layers, generate_text
